Module Asm.Ast

Copyright 2021-2022, Startsev Matvey

SPDX-License-Identifier: LGPL-3.0-or-later

type label =

| Label of string

label to use in code section such myLabel: and to use as argument of jmp commands

val pp_label : Ppx_deriving_runtime.Format.formatter -> label -> Ppx_deriving_runtime.unit

val show_label : label -> Ppx_deriving_runtime.string

type expr =

| Add of expr * expr

| Sub of expr * expr

| Mul of expr * expr

| Div of expr * expr

| Const of string	(*
interger constant
*)

| Var of string	(*
global constant
*)

| Reg of string	(*
comman rigister
*)

| Lab of label	(*
labels for jmps
*)

arithmetic expression or register to use as argument

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

type mnemonic =

| Mnemonic of string

just name of command

val pp_mnemonic : Ppx_deriving_runtime.Format.formatter -> mnemonic -> Ppx_deriving_runtime.unit

val show_mnemonic : mnemonic -> Ppx_deriving_runtime.string

type command =

| Args0 of mnemonic

| Args1 of mnemonic * expr

| Args2 of mnemonic * expr * expr

command with her arguments

val pp_command : Ppx_deriving_runtime.Format.formatter -> command -> Ppx_deriving_runtime.unit

val show_command : command -> Ppx_deriving_runtime.string

type code_section =

| Command of command

| Id of label

one line of code incode section that contains command and her arguments or label to jump

val pp_code_section : Ppx_deriving_runtime.Format.formatter -> code_section -> Ppx_deriving_runtime.unit

val show_code_section : code_section -> Ppx_deriving_runtime.string

type data_type =

| DataType of string

data type to use in data section

val pp_data_type : Ppx_deriving_runtime.Format.formatter -> data_type -> Ppx_deriving_runtime.unit

val show_data_type : data_type -> Ppx_deriving_runtime.string

type var =

| Variable of string * data_type * string list

one line of data declaration in data section that containd name of constant her type and value

val pp_var : Ppx_deriving_runtime.Format.formatter -> var -> Ppx_deriving_runtime.unit

val show_var : var -> Ppx_deriving_runtime.string

type dir =

| Code of code_section list

| Data of var list

code and data sections with

val pp_dir : Ppx_deriving_runtime.Format.formatter -> dir -> Ppx_deriving_runtime.unit

val show_dir : dir -> Ppx_deriving_runtime.string

type ast = dir list

main ast

val pp_ast : Ppx_deriving_runtime.Format.formatter -> ast -> Ppx_deriving_runtime.unit

val show_ast : ast -> Ppx_deriving_runtime.string