Module Asm_lib.Ast

Mini Assembler AST.

type id =

| Id of string

id contains an assembler label.

val pp_id : Ppx_deriving_runtime.Format.formatter -> id -> Ppx_deriving_runtime.unit

val show_id : id -> Ppx_deriving_runtime.string

type mnemonic =

| Mnemonic of string

mnemonic contains the name of the commands.

val pp_mnemonic : Ppx_deriving_runtime.Format.formatter -> mnemonic -> Ppx_deriving_runtime.unit

val show_mnemonic : mnemonic -> Ppx_deriving_runtime.string

type expr =

| Add of expr * expr

| Sub of expr * expr

| Mul of expr * expr

| Div of expr * expr

| Mod of expr * expr

| Shl of expr * expr

| Shr of expr * expr

| Reg of string

| Label of string

| Const of int64

expr contains the expression structure.

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

type init_value =

| Expr of expr

| Str of string

| Dup of expr * string

init_value contains the initial value of the expression or string.

val pp_init_value : Ppx_deriving_runtime.Format.formatter -> init_value -> Ppx_deriving_runtime.unit

val show_init_value : init_value -> Ppx_deriving_runtime.string

type command =

| Arg0 of mnemonic

| Arg1 of mnemonic * expr

| Arg2 of mnemonic * expr * expr

command contains the assembler instruction and the allowed number of arguments.

val pp_command : Ppx_deriving_runtime.Format.formatter -> command -> Ppx_deriving_runtime.unit

val show_command : command -> Ppx_deriving_runtime.string

type in_sec_dir =

| Instr of id option * command

| DataDecl of id option * string * init_value list

| EquDir of id * expr

| EqualDir of id * expr

in_sec_dir divides the section into data and code segments.

val pp_in_sec_dir : Ppx_deriving_runtime.Format.formatter -> in_sec_dir -> Ppx_deriving_runtime.unit

val show_in_sec_dir : in_sec_dir -> Ppx_deriving_runtime.string

type sec_dir =

| Code of in_sec_dir list

| Data of in_sec_dir list

sec_dir splits directive into sections, which begin with the keyword SECTION and the section type .CODE, .TEXT, .DATA or .CONST.

val pp_sec_dir : Ppx_deriving_runtime.Format.formatter -> sec_dir -> Ppx_deriving_runtime.unit

val show_sec_dir : sec_dir -> Ppx_deriving_runtime.string

type directive =

| Directive of sec_dir list

The main type for assembler AST.

val pp_directive : Ppx_deriving_runtime.Format.formatter -> directive -> Ppx_deriving_runtime.unit

val show_directive : directive -> Ppx_deriving_runtime.string

type simpl_dir = (id option * command) list

simpl_dir contains a list of all commands for the interpreter. The list is generated at the interpreter.

val pp_simpl_dir : Ppx_deriving_runtime.Format.formatter -> simpl_dir -> Ppx_deriving_runtime.unit

val show_simpl_dir : simpl_dir -> Ppx_deriving_runtime.string