Module Interpret.Eval

Main interpreter module.

Parameters

module M : MonadError

Signature

type var_t =

| R64 of int64

| RSSE of string

| Ls of string

| Er

Container for variables.

val pp_var_t : Ppx_deriving_runtime.Format.formatter -> var_t -> Ppx_deriving_runtime.unit

val show_var_t : var_t -> Ppx_deriving_runtime.string

type env = var_t MapString.t

Enviroment variables.

val pp_env : Ppx_deriving_runtime.Format.formatter -> env -> Ppx_deriving_runtime.unit

val show_env : env -> Ppx_deriving_runtime.string

val modulo : int64 -> int64 -> int64

modulo x y returns the positive remainder of the Int64 division.

val string_to_int64 : string -> int64

string_to_int64 s takes a string s, interprets it as a bit little-endian number and returns it as Int64.

val int64_to_string : int64 -> string

int64_to_string i takes an integer i, interprets it as a hexadecimal number, makes and returns a little-endian string.

val ev : 'a -> (Ast.expr -> int64 M.t) -> Ast.expr -> int64 M.t

ev e f counts arithmetic operations, takes an expr and f that decides to process registers Ast.Reg and labels Ast.Label.

val prepr : var_t MapString.t -> Ast.directive -> (var_t MapString.t * Ast.command list) M.t

prepr env dir accepts the environment and Ast.Directive, returns the environment filled with the labels found and the simplified directory with the instructions Ast.simpl_dir.

val reg_list : (string * var_t) list

List of all registers.

val reg8 : string -> bool

val reg8L : string -> bool

val reg16 : string -> bool

val reg32 : string -> bool

val reg64 : string -> bool

val regSSE : string -> bool

val interpret : (var_t MapString.t * Ast.command list) -> var_t MapString.t M.t

interpret (env, list) executes commands from a list in this environment.

val asm : Ast.directive -> var_t MapString.t M.t

Main interpreter function.