Module Asm_lib.Parser

Mini Assembler Parser.

val spaces : unit Angstrom.t

spaces accepts input as long as it's space and discards the accepted characters.

val is_blank : char -> bool

is_blank c accepts any character and returns whether c is blank.

val eol : unit list Angstrom.t

eol accepts input as long as it's blank or comment and discards the accepted characters.

val trim : 'a Angstrom.t -> 'a Angstrom.t

trim x accepts input and discards space characters around.

val is_digit : char -> bool

is_digit accepts any character and returns whether it's digit.

val is_register : string -> bool

is_register accepts string and returns whether it's register.

val is_arg0 : string -> bool

is_arg0 accepts assembler command, returns true if the command expects no arguments.

val is_arg1 : string -> bool

is_arg1 accepts assembler command, returns true if command takes 1 argument.

val is_arg2 : string -> bool

is_arg2 accepts assembler command, returns true if command takes 2 arguments.

val is_mnemonic : string -> bool

is_mnemonic accepts string and returns whether it's assembler instruction.

val is_section : string -> bool

is_section accepts string and returns whether it's assembler section name.

val is_data_decl : string -> bool

is_data_decl accepts string and returns whether it's assembler data type.

val take_id : string Angstrom.t

take_id accepts input returns string if it is assembler label or register otherwise fails.

val expr_p : Ast.expr Angstrom.t

expr_p parses Ast.Add, Ast.Sub, Ast.Mul, Ast.Div, Ast.Mod, Ast.Shl, Ast.Shr, Ast.Reg, Ast.Label and Ast.Const.

val init_value_p : Ast.init_value Angstrom.t

init_value_p parses Ast.Expr, Ast.Str and Ast.Dup.

val mnemonic_p : Ast.mnemonic Angstrom.t

mnemonic_p parses Ast.Mnemonic.

val in_sec_dir_p : Ast.in_sec_dir Angstrom.t

val sec_dir_p : Ast.sec_dir Angstrom.t

sec_dir_p parses Ast.Section.

val directive_p : Ast.directive Angstrom.t

directive_p parses Ast.Directive.

val parse : string -> (Ast.directive, string) Stdlib.result

parse runs directive_p on s. The parser will receive an `Eof after all of s has been consumed.

Helper functions for testing.

val parse_test : 'a Angstrom.t -> string -> ('a, string) Stdlib.result

val parse_test_suc : (Stdlib.Format.formatter -> 'a -> unit) -> 'a Angstrom.t -> string -> 'a -> bool

val parse_test_fail : (Stdlib.Format.formatter -> 'a -> unit) -> 'a Angstrom.t -> string -> bool

val expr_test_suc : string -> Ast.expr -> bool

val expr_test_fail : string -> bool

val init_value_test_suc : string -> Ast.init_value -> bool

val init_value_test_fail : string -> bool

val in_sec_dir_test_suc : string -> Ast.in_sec_dir -> bool

val in_sec_dir_test_fail : string -> bool

val sec_dir_test_suc : string -> Ast.sec_dir -> bool

val sec_dir_test_fail : string -> bool

val directive_suc : string -> Ast.directive -> bool