Module Ast

Copyright 2023-2024, Vadim Yakshigulov

SPDX-License-Identifier: LGPL-3.0-or-later

type i64

type i128

type _ register =

| Rax : i64 register

| Rbx : i64 register

| Rcx : i64 register

| Rdx : i64 register

| Rsp : i64 register

| Rbp : i64 register

| Rsi : i64 register

| Rdi : i64 register

| Xmm0 : i128 register

| Xmm1 : i128 register

| Xmm2 : i128 register

| Xmm3 : i128 register

| Xmm4 : i128 register

| Xmm5 : i128 register

| Xmm6 : i128 register

| Xmm7 : i128 register

val rax : i64 register

val rbx : i64 register

val rcx : i64 register

val rdx : i64 register

val rsp : i64 register

val rbp : i64 register

val rsi : i64 register

val rdi : i64 register

val xmm0 : i128 register

val xmm1 : i128 register

val xmm2 : i128 register

val xmm3 : i128 register

val xmm4 : i128 register

val xmm5 : i128 register

val xmm6 : i128 register

val xmm7 : i128 register

val is_rax : 'a register -> bool

val is_rbx : 'a register -> bool

val is_rcx : 'a register -> bool

val is_rdx : 'a register -> bool

val is_rsp : 'a register -> bool

val is_rbp : 'a register -> bool

val is_rsi : 'a register -> bool

val is_rdi : 'a register -> bool

val is_xmm0 : 'a register -> bool

val is_xmm1 : 'a register -> bool

val is_xmm2 : 'a register -> bool

val is_xmm3 : 'a register -> bool

val is_xmm4 : 'a register -> bool

val is_xmm5 : 'a register -> bool

val is_xmm6 : 'a register -> bool

val is_xmm7 : 'a register -> bool

module Variants_of_register : sig ... end

type _ immediate =

| Imm_int of int

val imm_int : int -> 'a immediate

val is_imm_int : 'a immediate -> bool

val imm_int_val : 'a immediate -> int Stdlib.Option.t

module Variants_of_immediate : sig ... end

type 'a register_ref =

| Register_ref of 'a register

val register_ref : 'a register -> 'a register_ref

val is_register_ref : 'a register_ref -> bool

val register_ref_val : 'a register_ref -> 'a register Stdlib.Option.t

module Variants_of_register_ref : sig ... end

type _ operand =

| Reg_64_reg_64 : i64 register * i64 register -> [> `Reg_64_reg_64 ] operand
(*
Double register operands
*)

| Mem_64_mem_64 : i64 register_ref * i64 register_ref -> [> `Mem_64_mem_64 ] operand
(*
Double memory operands
*)

| Reg_64_imm_a : i64 register * _ immediate -> [> `Reg_64_imm_a ] operand
(*
Mixed double operands
*)

| Imm_a : _ immediate -> [> `Imm_a ] operand
(*
Single immidiate operand
*)

| Reg_64 : i64 register -> [> `Reg_64 ] operand
(*
Single register operand
*)

| Mem_64 : i64 register_ref -> [> `Mem_64 ] operand
(*
Single memory operand
*)

| Nothing : [> `Nothing ] operand
(*
Mnemonic without operand (e.g Ret)
*)

| Label : string -> [> `Label ] operand
(*
Label operand (e.g Jmp lbl)
*)

| Reg_128 : i128 register -> [> `Reg_128 ] operand

| Reg_64_reg_128 : i64 register * i128 register -> [> `Reg_64_reg_128 ] operand

| Reg_128_reg_64 : i128 register * i64 register -> [> `Reg_128_reg_64 ] operand

| Reg_128_reg_128 : i128 register * i128 register -> [> `Reg_128_reg_128 ] operand

| Reg_128_reg_64_imm_a : i128 register * i64 register * 'a immediate -> [> `Reg_128_reg_64_imm_a ] operand

val reg_64_reg_64 : i64 register -> i64 register -> [> `Reg_64_reg_64 ] operand

val mem_64_mem_64 : i64 register_ref -> i64 register_ref -> [> `Mem_64_mem_64 ] operand

val reg_64_imm_a : i64 register -> 'a immediate -> [> `Reg_64_imm_a ] operand

val imm_a : 'a immediate -> [> `Imm_a ] operand

val reg_64 : i64 register -> [> `Reg_64 ] operand

val mem_64 : i64 register_ref -> [> `Mem_64 ] operand

val nothing : [> `Nothing ] operand

val label : string -> [> `Label ] operand

val reg_128 : i128 register -> [> `Reg_128 ] operand

val reg_64_reg_128 : i64 register -> i128 register -> [> `Reg_64_reg_128 ] operand

val reg_128_reg_64 : i128 register -> i64 register -> [> `Reg_128_reg_64 ] operand

val reg_128_reg_128 : i128 register -> i128 register -> [> `Reg_128_reg_128 ] operand

val reg_128_reg_64_imm_a : i128 register -> i64 register -> 'a immediate -> [> `Reg_128_reg_64_imm_a ] operand

val is_reg_64_reg_64 : 'a operand -> bool

val is_mem_64_mem_64 : 'a operand -> bool

val is_reg_64_imm_a : 'a operand -> bool

val is_imm_a : 'a operand -> bool

val is_reg_64 : 'a operand -> bool

val is_mem_64 : 'a operand -> bool

val is_nothing : 'a operand -> bool

val is_label : 'a operand -> bool

val is_reg_128 : 'a operand -> bool

val is_reg_64_reg_128 : 'a operand -> bool

val is_reg_128_reg_64 : 'a operand -> bool

val is_reg_128_reg_128 : 'a operand -> bool

val is_reg_128_reg_64_imm_a : 'a operand -> bool

module Variants_of_operand : sig ... end

type _ mnemonic =

| Mov : [< `Reg_64_imm_a ] mnemonic

| Push : [< `Reg_64 ] mnemonic

| Add : [< `Reg_64_reg_64 ] mnemonic

| Xor : [< `Reg_64_reg_64 ] mnemonic

| Syscall : [< `Nothing ] mnemonic

| Pop : [< `Reg_64 ] mnemonic

| Jmp : [< `Label ] mnemonic

| Sub : [< `Reg_64_imm_a ] mnemonic

| Cmp : [< `Reg_64_imm_a ] mnemonic

| Je : [< `Label ] mnemonic

| Jne : [< `Label ] mnemonic

| Jle : [< `Label ] mnemonic

| Movq : [< `Reg_64_reg_128 | `Reg_128_reg_64 ] mnemonic

| Punpckhqdq : [< `Reg_128_reg_128 ] mnemonic

| Movapd : [< `Reg_128_reg_128 ] mnemonic
(*
Real instruction name, not the cat jumped on the keyboard
*)

| Pinsrq : [< `Reg_128_reg_64_imm_a ] mnemonic

| Addpd : [< `Reg_128_reg_128 ] mnemonic

| Mulpd : [< `Reg_128_reg_128 ] mnemonic

| Haddpd : [< `Reg_128_reg_128 ] mnemonic
(*
Not sure that these instructions can really be used with integer values in xmm registers. The specification of the instructions is a complete mess
*)

Some (actually most of them) mnemonic variants are not supported, but it's easy to add them later

val mov : [< `Reg_64_imm_a ] mnemonic

val push : [< `Reg_64 ] mnemonic

val add : [< `Reg_64_reg_64 ] mnemonic

val xor : [< `Reg_64_reg_64 ] mnemonic

val syscall : [< `Nothing ] mnemonic

val pop : [< `Reg_64 ] mnemonic

val jmp : [< `Label ] mnemonic

val sub : [< `Reg_64_imm_a ] mnemonic

val cmp : [< `Reg_64_imm_a ] mnemonic

val je : [< `Label ] mnemonic

val jne : [< `Label ] mnemonic

val jle : [< `Label ] mnemonic

val movq : [< `Reg_128_reg_64 | `Reg_64_reg_128 ] mnemonic

val punpckhqdq : [< `Reg_128_reg_128 ] mnemonic

val movapd : [< `Reg_128_reg_128 ] mnemonic

val pinsrq : [< `Reg_128_reg_64_imm_a ] mnemonic

val addpd : [< `Reg_128_reg_128 ] mnemonic

val mulpd : [< `Reg_128_reg_128 ] mnemonic

val haddpd : [< `Reg_128_reg_128 ] mnemonic

val is_mov : 'a mnemonic -> bool

val is_push : 'a mnemonic -> bool

val is_add : 'a mnemonic -> bool

val is_xor : 'a mnemonic -> bool

val is_syscall : 'a mnemonic -> bool

val is_pop : 'a mnemonic -> bool

val is_jmp : 'a mnemonic -> bool

val is_sub : 'a mnemonic -> bool

val is_cmp : 'a mnemonic -> bool

val is_je : 'a mnemonic -> bool

val is_jne : 'a mnemonic -> bool

val is_jle : 'a mnemonic -> bool

val is_movq : 'a mnemonic -> bool

val is_punpckhqdq : 'a mnemonic -> bool

val is_movapd : 'a mnemonic -> bool

val is_pinsrq : 'a mnemonic -> bool

val is_addpd : 'a mnemonic -> bool

val is_mulpd : 'a mnemonic -> bool

val is_haddpd : 'a mnemonic -> bool

module Variants_of_mnemonic : sig ... end

type directive =

| Section of string

| Global of string

val section : string -> directive

val global : string -> directive

val is_section : directive -> bool

val is_global : directive -> bool

val section_val : directive -> string Stdlib.Option.t

val global_val : directive -> string Stdlib.Option.t

module Variants_of_directive : sig ... end

type statement =

| Directive of directive

| Label_decl of string

| Instruction : 'kind_of mnemonic * 'kind_of operand -> statement

val directive : directive -> statement

val label_decl : string -> statement

val instruction : 'a mnemonic -> 'a operand -> statement

val is_directive : statement -> bool

val is_label_decl : statement -> bool

val is_instruction : statement -> bool

val directive_val : statement -> directive Stdlib.Option.t

val label_decl_val : statement -> string Stdlib.Option.t

module Variants_of_statement : sig ... end

type ast = statement list

val show_register : 'a register -> string

val show_mnemonic : 'a mnemonic -> string

val show_immediate : 'a immediate -> string

val show_register_ref : 'a register_ref -> string

val show_directive : directive -> string

type args_wrapper =

| Wargs : 'kind_of operand -> args_wrapper

val wargs : 'a operand -> args_wrapper

val is_wargs : args_wrapper -> bool

module Variants_of_args_wrapper : sig ... end

type register_wrapper =

| Wregister : 'a register -> register_wrapper

val wregister : 'a register -> register_wrapper

val is_wregister : register_wrapper -> bool

module Variants_of_register_wrapper : sig ... end

val show_operand : args_wrapper -> string

val show_statement : statement -> string

val show_ast : statement list -> string